# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.FileResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      expires_at:
        example: "2024-02-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      mime_type:
        example: application/pdf
        type: string
      name:
        example: document.pdf
        type: string
      size:
        example: 1024000
        type: integer
      storage_path:
        example: users/550e8400-e29b-41d4-a716-446655440000/files/...
        type: string
      storage_type:
        example: local
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handlers.ListFilesResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/handlers.FileResponse'
        type: array
      total:
        example: 10
        type: integer
    type: object
...
  handlers.TurnCredentialsResponse:
    properties:
      password:
        type: string
      realm:
        type: string
      turn_servers:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
...
  handlers.UploadFileResponse:
    properties:
      file_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      storage_path:
        example: users/550e8400-e29b-41d4-a716-446655440000/files/...
        type: string
      uploaded_size:
        example: 1024000
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@Backend.app
    name: Backend Support
    url: https://Backend.app/support
  description: Backend - SaaS приложение для файлообмена с акцентом на максимальную скорость
    передачи через Local Network First подход
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://Backend.app/terms
  title: Backend API
  version: "1.0"
paths:
...
  /files:
    get:
      consumes:
      - application/json
      description: Возвращает список файлов пользователя с пагинацией
      parameters:
      - default: 50
        description: Лимит файлов
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список файлов
          schema:
            $ref: '#/definitions/handlers.ListFilesResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список файлов
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл на сервер через multipart/form-data. Поддерживает
        потоковую загрузку больших файлов.
      parameters:
      - description: Файл для загрузки
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Файл успешно загружен
          schema:
            $ref: '#/definitions/handlers.UploadFileResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузка файла
      tags:
      - files
  /files/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет файл с сервера
      parameters:
      - description: ID файла
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Файл удален
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к файлу
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Файл не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление файла
      tags:
      - files
    get:
      consumes:
      - application/json
      description: Возвращает метаданные файла
      parameters:
      - description: ID файла
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Метаданные файла
          schema:
            $ref: '#/definitions/handlers.FileResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к файлу
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Файл не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Метаданные файла
      tags:
      - files
  /files/{id}/download:
    get:
      consumes:
      - application/json
      description: Скачивает файл с сервера (поддерживает Range requests для частичного
        скачивания). Использует потоковую передачу для больших файлов.
      parameters:
      - description: ID файла
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Смещение в байтах для частичного скачивания
        in: query
        name: offset
        type: integer
      - default: 0
        description: Лимит байт для чтения (0 = весь файл)
        in: query
        name: limit
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Файл (бинарные данные)
          schema:
            type: file
        "400":
          description: Неверный ID файла" example:"{\"error\":\"file_id is required\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован" example:"{\"error\":\"unauthorized\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к файлу" example:"{\"error\":\"permission denied\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Файл не найден" example:"{\"error\":\"file not found\"}
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера" example:"{\"error\":\"failed to
            download file\"}
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Скачивание файла
      tags:
      - files
  /webrtc/turn-credentials:
    get:
      consumes:
      - application/json
      description: Возвращает TURN серверы и credentials для WebRTC
      produces:
      - application/json
      responses:
        "200":
          description: TURN credentials
          schema:
            $ref: '#/definitions/handlers.TurnCredentialsResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение TURN credentials
      tags:
      - webrtc
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"